refactor: variant header and summary
