+++ src/pages/VariantPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './VariantPage';
