+++ src/queries/VariantPageQuery.gql
@@ -1,36 +1,4 @@
 query VariantPageQuery($variantId: String!) {
-  variantInfo(variantId: $variantId) {
-    rsId
-    chromosome
-    position
-    chromosomeB37
-    positionB37
-    refAllele
-    altAllele
-    nearestGene {
-      id
-      symbol
-    }
-    nearestGeneDistance
-    nearestCodingGene {
-      id
-      symbol
-    }
-    nearestCodingGeneDistance
-    mostSevereConsequence
-    caddRaw
-    caddPhred
-    gnomadAFR
-    gnomadAMR
-    gnomadASJ
-    gnomadEAS
-    gnomadFIN
-    gnomadNFE
-    gnomadNFEEST
-    gnomadNFENWE
-    gnomadNFESEU
-    gnomadOTH
-  }
   genesForVariantSchema {
     qtls {
       id
